move linearupsample and makecontiguous to nn
